requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.8.3
python-dotenv==1.0.1
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.1.0
yfinance==0.2.28
dnspython==2.4.2 